Update Change Role Flow

+++ backend/api-gateway/handlers/user.go
@@ -13,10 +13,37 @@ import (
 	"gorm.io/gorm"
 )
 
+func GetUsers(c echo.Context) error {
+	tokenClaims := c.Get(TOKEN_CLAIMS_CONTEXT_KEY).(*models.Claims)
+
+	superAdmin := tokenClaims.User
+
+	if superAdmin.Role != SUPER_ADMIN {
+		return c.JSON(http.StatusForbidden, message.CreateErrorMessage(FAILURE_NOT_SUPERADMIN_GET_USERS))
+	}
+
+	users := make([]models.User, 0)
+	if err := config.DB.Find(&users).Error; err != nil {
+		return c.JSON(http.StatusInternalServerError, message.CreateErrorMessage(INVALID_DB_ERROR))
+	}
+	return c.JSON(http.StatusOK, message.CreateSuccessUsersMessage(SUCCESS_USER_FOUND, users))
+}
+
 func GetUser(c echo.Context) error {
 	tokenClaims := c.Get(TOKEN_CLAIMS_CONTEXT_KEY).(*models.Claims)
 
-	user := tokenClaims.User
+	currentUser := tokenClaims.User
+	oauthId := currentUser.OauthID
+	oauthProvider := currentUser.OauthProvider
+
+	var user models.User
+	err := config.DB.Where("oauth_id = ? AND oauth_provider = ?", oauthId, oauthProvider).First(&user).Error
+	if err != nil {
+		if err == gorm.ErrRecordNotFound {
+			return c.JSON(http.StatusBadRequest, message.CreateErrorMessage(INVALID_USER_NOT_FOUND))
+		}
+		return c.JSON(http.StatusInternalServerError, message.CreateErrorMessage(INVALID_DB_ERROR))
+	}
 
 	return c.JSON(http.StatusOK, message.CreateSuccessUserMessage(SUCCESS_USER_FOUND, user))
 }
@@ -25,6 +52,10 @@ func DeleteUser(c echo.Context) error {
 	tokenClaims := c.Get(TOKEN_CLAIMS_CONTEXT_KEY).(*models.Claims)
 
 	user := tokenClaims.User
+
+	if user.Role == SUPER_ADMIN {
+		return c.JSON(http.StatusForbidden, message.CreateErrorMessage(FAILURE_DELETE_SUPERADMIN))
+	}
 
 	authUserId := user.ID
 	responseStatusCode, responseMessage := client.UserService.DeleteUser(authUserId)
